+++ src/mplchart/primitives/__init__.py
@@ -6,3 +6,8 @@ from .volume import Volume
 from .ohlc import OHLC
 from .candlesticks import Candlesticks
 from .peaks import Peaks
+from .lineplot import LinePlot
+from .areaplot import AreaPlot
+from .barplot import BarPlot
+
+