+++ src/mplchart/primitives/ohlc.py
@@ -24,15 +24,15 @@ class OHLC(Primitive):
     def __str__(self):
         return self.__class__.__name__
 
-    def plot_handler(self, data, chart, ax=None):
+    def plot_handler(self, prices, chart, ax=None):
         if ax is None:
             ax = chart.get_axes()
 
-        label = str(self)
-        data = chart.extract_df(data)
+        data = chart.extract_df(prices)
 
         edgecolor = plt.rcParams["text.color"]
 
+        label = str(self)
         width = self.width
         colorup = self.colorup or edgecolor
         colordn = self.colordn or edgecolor