+++ src/mplchart/primitives/price.py
@@ -1,9 +1,7 @@
 """ Price primitive """
 
-import pandas as pd
-
 from ..model import Primitive
-from ..utils import series_xy
+from ..utils import series_data, series_xy
 
 
 class Price(Primitive):
@@ -15,36 +13,24 @@ class Price(Primitive):
     Args:
         item (str) :  name of the column to plot. default 'close'
 
-    Returns:
-        the series of given name
-
     Example:
-        Price('close') the close price series
-        Price('open') the open price series
-
+        Price('close')  # plot the close price series
+        Price('open')   # plot the open price series
     """
 
-    def __init__(self, item: str = "close", *, color: str = None):
+    def __init__(self, item: str = 'close', *, color: str = None):
         self.item = item
         self.color = color
 
-    def __str__(self):
-        return self.item
-
-    def calc(self, data):
-        if isinstance(data, pd.DataFrame):
-            data = data[self.item]
-        return data
-
-    def plot_handler(self, data, chart, ax=None):
+    def plot_handler(self, prices, chart, ax=None):
         if ax is None:
             ax = chart.get_axes()
 
-        label = str(self)
-        color = self.color
-
-        data = self.calc(data)
+        data = series_data(prices, self.item)
         data = chart.extract_df(data)
+
+        label = self.item
+        color = self.color
 
         xv, yv = series_xy(data)
         ax.plot(xv, yv, label=label, color=color)